+++ packages/components/button/src/index.ts
@@ -1,2 +1,4 @@
+export { ButtonGroup } from './ButtonGroup';
+export type { ButtonGroupProps } from './ButtonGroup';
 export { Button } from './Button';
 export type { ButtonProps } from './Button';
