+++ packages/forma-36-react-components/src/index.ts
@@ -69,8 +69,8 @@ export type {
 } from '@contentful/f36-inputs';
 export { Pill } from '@contentful/f36-pill';
 export type { PillProps } from '@contentful/f36-pill';
-export { ButtonGroup } from '@contentful/f36-buttongroup';
-export type { ButtonGroupProps } from '@contentful/f36-buttongroup';
+export { ButtonGroup } from '@contentful/f36-button';
+export type { ButtonGroupProps } from '@contentful/f36-button';
 export { Button } from '@contentful/f36-button';
 export type { ButtonProps } from '@contentful/f36-button';
 export * from './components/Accordion';
